python/confluence/pyproject.toml
[project]
name = "confluence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
    "truststore>=0.10.4",
]
